@@ -34,39 +34,41 @@
   "dependencies": {
     "blacklist": "^1.1.4",
     "dompurify": "^0.8.4",
-    "is-dom": "^1.0.5"
+    "is-dom": "^1.0.5",
+    "prop-types": "latest"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.21.0",
-    "babel-loader": "^6.2.10",
-    "babel-polyfill": "^6.20.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-power-assert": "^1.0.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-2": "^6.18.0",
-    "cssstyle": "^0.2.37",
-    "deep-equal": "^1.0.1",
+    "babel-cli": "~6.26.0",
+    "babel-core": "~6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-polyfill": "~6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-power-assert": "^2.0.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "cssstyle": "^2.0.0",
+    "deep-equal": "^1.1.0",
     "espower-babel": "^4.0.3",
     "json-loader": "^0.5.4",
-    "karma": "^1.3.0",
+    "karma": "^4.3.0",
     "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.1",
-    "karma-phantomjs-launcher": "^1.0.2",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-phantomjs-launcher": "^1.0.4",
     "karma-sinon": "^1.0.5",
-    "karma-webpack": "^2.0.1",
-    "mocha": "^3.2.0",
-    "phantomjs-prebuilt": "^2.1.14",
-    "power-assert": "^1.4.2",
-    "react": "^15.4.1",
-    "react-addons-test-utils": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-redux": "^5.1.2",
-    "redux": "^3.6.0",
-    "redux-logger": "^2.7.4",
-    "rimraf": "^2.5.4",
-    "sinon": "^1.17.6",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2"
+    "karma-webpack": "^4.0.2",
+    "mocha": "^6.2.1",
+    "phantomjs-prebuilt": "^2.1.16",
+    "power-assert": "^1.6.1",
+    "react": "^16.10.2",
+    "react-addons-test-utils": "^16.0.0-alpha.3",
+    "react-dom": "^16.10.2",
+    "react-redux": "^7.1.1",
+    "redux": "^4.0.4",
+    "redux-logger": "^4.0.0",
+    "rimraf": "^3.0.0",
+    "sinon": "^7.5.0",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.8.2"
   }
 }
